+++ src/entity/importBillStock.ts
@@ -24,6 +24,6 @@ export class ImportBillStock extends BaseEntity {
     @Column('smallint', { default: 1 })
     public quantity: number;
 
-    @Column('money', { nullable: true })
+    @Column('float', { nullable: true })
     public price: number;
 }
